RF(BM): extract settingup of a sample repo into SampleSuperDatasetBenchmarks class
